Extract MiscV1SecurityModelSuite
