+++ hedera-node/hedera-smart-contract-service-impl/src/test/java/com/hedera/node/app/service/contract/impl/test/exec/systemcontracts/hts/transfer/Erc20TransfersCallTest.java
@@ -77,7 +77,8 @@ class Erc20TransfersCallTest extends HtsCallTestBase {
                 null,
                 verificationStrategy,
                 SENDER_ID,
-                addressIdConverter);
+                addressIdConverter,
+                false);
 
         final var result = subject.execute().fullResult().result();
 
@@ -161,7 +162,8 @@ class Erc20TransfersCallTest extends HtsCallTestBase {
                 FUNGIBLE_TOKEN_ID,
                 verificationStrategy,
                 SENDER_ID,
-                addressIdConverter);
+                addressIdConverter,
+                false);
     }
 
     private Erc20TransfersCall subjectForTransferFrom(final long amount) {
@@ -174,6 +176,7 @@ class Erc20TransfersCallTest extends HtsCallTestBase {
                 FUNGIBLE_TOKEN_ID,
                 verificationStrategy,
                 SENDER_ID,
-                addressIdConverter);
+                addressIdConverter,
+                false);
     }
 }
