+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/precompile/ApproveAllowanceSuite.java
@@ -123,6 +123,7 @@ public class ApproveAllowanceSuite extends HapiSuite {
                 testIndirectApprovalWithDirectErc20Callee());
     }
 
+    @HapiTest
     private HapiSpec nftAutoCreationIncludeAllowanceCheck() {
         final var ownerAccount = "owningAlias";
         final var receivingAlias = "receivingAlias";
