+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/precompile/AtomicCryptoTransferHTSSuite.java
@@ -17,7 +17,7 @@
 package com.hedera.services.bdd.suites.contract.precompile;
 
 import static com.hedera.services.bdd.junit.TestTags.SMART_CONTRACT;
-import static com.hedera.services.bdd.spec.HapiSpec.defaultHapiSpec;
+import static com.hedera.services.bdd.spec.HapiSpec.propertyPreservingHapiSpec;
 import static com.hedera.services.bdd.spec.assertions.AccountDetailsAsserts.accountDetailsWith;
 import static com.hedera.services.bdd.spec.assertions.ContractFnResultAsserts.resultWith;
 import static com.hedera.services.bdd.spec.assertions.TransactionRecordAsserts.recordWith;
@@ -49,6 +49,7 @@ import static com.hedera.services.bdd.spec.utilops.UtilVerbs.childRecordsCheck;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.newKeyNamed;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.nftTransfer;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overriding;
+import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overridingTwo;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.tokenTransferList;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.tokenTransferLists;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.transferList;
@@ -66,6 +67,7 @@ import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.SUCCESS;
 
 import com.esaulpaugh.headlong.abi.Tuple;
 import com.hedera.node.app.hapi.utils.ByteStringUtils;
+import com.hedera.services.bdd.junit.HapiTest;
 import com.hedera.services.bdd.junit.HapiTestSuite;
 import com.hedera.services.bdd.spec.HapiSpec;
 import com.hedera.services.bdd.spec.assertions.ContractInfoAsserts;
@@ -114,7 +116,7 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
 
     @Override
     public boolean canRunConcurrent() {
-        return true;
+        return false;
     }
 
     @Override
@@ -131,6 +133,7 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
         });
     }
 
+    @HapiTest
     private HapiSpec cryptoTransferForHbarOnly() {
         final var cryptoTransferTxn = "cryptoTransferTxn";
         final var cryptoTransferMultiTxn = "cryptoTransferMultiTxn";
@@ -138,9 +141,12 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
         final var cryptoTransferRevertNoKeyTxn = "cryptoTransferRevertNoKeyTxn";
         final var cryptoTransferRevertBalanceTooLowTxn = "cryptoTransferRevertBalanceTooLowTxn";
 
-        return defaultHapiSpec("cryptoTransferForHbarOnly")
+        return propertyPreservingHapiSpec("cryptoTransferForHbarOnly")
+                .preserving("contracts.allowAutoAssociations", "contracts.precompile.atomicCryptoTransfer.enabled")
                 .given(
-                        overriding("contracts.allowAutoAssociations", "true"),
+                        overridingTwo(
+                                "contracts.allowAutoAssociations", "true",
+                                "contracts.precompile.atomicCryptoTransfer.enabled", "true"),
                         cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
                         cryptoCreate(SENDER2).balance(10 * ONE_HUNDRED_HBARS),
                         cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
@@ -321,13 +327,16 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
                                                         htsPrecompileResult().withStatus(INVALID_ACCOUNT_AMOUNTS)))));
     }
 
+    @HapiTest
     private HapiSpec cryptoTransferForFungibleTokenOnly() {
-
         final var cryptoTransferTxnForFungible = "cryptoTransferTxnForFungible";
 
-        return defaultHapiSpec("cryptoTransferForFungibleTokenOnly")
+        return propertyPreservingHapiSpec("cryptoTransferForFungibleTokenOnly")
+                .preserving("contracts.allowAutoAssociations", "contracts.precompile.atomicCryptoTransfer.enabled")
                 .given(
-                        overriding("contracts.allowAutoAssociations", "true"),
+                        overridingTwo(
+                                "contracts.allowAutoAssociations", "true",
+                                "contracts.precompile.atomicCryptoTransfer.enabled", "true"),
                         cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
                         cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
                         cryptoCreate(TOKEN_TREASURY),
@@ -393,11 +402,14 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
                                                 .including(FUNGIBLE_TOKEN, RECEIVER, 50))));
     }
 
+    @HapiTest
     private HapiSpec cryptoTransferForNonFungibleTokenOnly() {
         final var cryptoTransferTxnForNft = "cryptoTransferTxnForNft";
 
-        return defaultHapiSpec("cryptoTransferForNonFungibleTokenOnly")
+        return propertyPreservingHapiSpec("cryptoTransferForNonFungibleTokenOnly")
+                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
                 .given(
+                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
                         newKeyNamed(MULTI_KEY),
                         cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
                         cryptoCreate(RECEIVER).receiverSigRequired(true),
@@ -464,11 +476,14 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
                                                 .including(NFT_TOKEN, SENDER, RECEIVER, 1L))));
     }
 
+    @HapiTest
     private HapiSpec cryptoTransferHBarFungibleNft() {
         final var cryptoTransferTxnForAll = "cryptoTransferTxnForAll";
 
-        return defaultHapiSpec("cryptoTransferHBarFungibleNft")
+        return propertyPreservingHapiSpec("cryptoTransferHBarFungibleNft")
+                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
                 .given(
+                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
                         newKeyNamed(MULTI_KEY),
                         cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
                         cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
@@ -589,6 +604,7 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
                                         .transfers(including(tinyBarsFromTo(SENDER, RECEIVER, 50 * ONE_HBAR)))));
     }
 
+    @HapiTest
     private HapiSpec cryptoTransferAllowanceHbarToken() {
         final var allowance = 10L;
         final var successfulTransferFromTxn = "txn";
@@ -597,8 +613,10 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
         final var revertingTransferFromTxn2 = "revertingTxn";
         final var revertingTransferFromTxn3 = "revertingTxnNoApproval";
 
-        return defaultHapiSpec("cryptoTransferAllowanceHbarToken")
+        return propertyPreservingHapiSpec("cryptoTransferAllowanceHbarToken")
+                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
                 .given(
+                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
                         newKeyNamed(MULTI_KEY),
                         cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS).maxAutomaticTokenAssociations(5),
                         cryptoCreate(RECEIVER).maxAutomaticTokenAssociations(5),
@@ -749,6 +767,7 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
                                                         .withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
     }
 
+    @HapiTest
     private HapiSpec cryptoTransferAllowanceFungibleToken() {
         final var allowance = 10L;
         final var successfulTransferFromTxn = "txn";
@@ -756,8 +775,10 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
         final var revertingTransferFromTxnFungible = "revertWhenMoreThanAllowanceFungible";
         final var revertingTransferFromTxn2 = "revertingTxn";
         final var revertingTransferFromTxn3 = "revertingTxnNoApproval";
-        return defaultHapiSpec("CryptoTransferAllowanceFungibleToken")
+        return propertyPreservingHapiSpec("CryptoTransferAllowanceFungibleToken")
+                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
                 .given(
+                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
                         newKeyNamed(MULTI_KEY),
                         cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS).maxAutomaticTokenAssociations(5),
                         cryptoCreate(RECEIVER).maxAutomaticTokenAssociations(5),
@@ -934,11 +955,14 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
                                                         .withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
     }
 
+    @HapiTest
     private HapiSpec cryptoTransferAllowanceNft() {
         final var successfulTransferFromTxn = "txn";
         final var revertingTransferFromTxnNft = "revertWhenMoreThanAllowanceNft";
-        return defaultHapiSpec("cryptoTransferAllowanceNft")
+        return propertyPreservingHapiSpec("cryptoTransferAllowanceNft")
+                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
                 .given(
+                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
                         newKeyNamed(MULTI_KEY),
                         cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS).maxAutomaticTokenAssociations(5),
                         cryptoCreate(RECEIVER).maxAutomaticTokenAssociations(5),
@@ -1015,11 +1039,16 @@ public class AtomicCryptoTransferHTSSuite extends HapiSuite {
                                                 .including(NFT_TOKEN, OWNER, RECEIVER, 2L))));
     }
 
+    // Expected INVALID_FULL_PREFIX_SIGNATURE_FOR_PRECOMPILE but was INVALID_ACCOUNT_ID
     private HapiSpec cryptoTransferSpecialAccounts() {
         final var cryptoTransferTxn = "cryptoTransferTxn";
 
-        return defaultHapiSpec("cryptoTransferEmptyKeyList")
+        return propertyPreservingHapiSpec("cryptoTransferEmptyKeyList")
+                .preserving("contracts.allowAutoAssociations", "contracts.precompile.atomicCryptoTransfer.enabled")
                 .given(
+                        overridingTwo(
+                                "contracts.allowAutoAssociations", "true",
+                                "contracts.precompile.atomicCryptoTransfer.enabled", "true"),
                         cryptoCreate(RECEIVER).balance(1 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
                         uploadInitCode(CONTRACT),
                         contractCreate(CONTRACT).maxAutomaticTokenAssociations(1),
