+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/precompile/ContractKeysStillWorkAsExpectedSuite.java
@@ -28,9 +28,7 @@ import static com.hedera.services.bdd.spec.keys.KeyShape.sigs;
 import static com.hedera.services.bdd.spec.keys.SigControl.ON;
 import static com.hedera.services.bdd.spec.keys.SigControl.SECP256K1_ON;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountBalance;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getContractInfo;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTokenInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTokenNftInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTxnRecord;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCall;
@@ -43,12 +41,6 @@ import static com.hedera.services.bdd.spec.transactions.TxnVerbs.mintToken;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenAssociate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.uploadInitCode;
-import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHbarFee;
-import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHbarFeeInheritingRoyaltyCollector;
-import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHtsFeeInheritingRoyaltyCollector;
-import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.royaltyFeeWithFallback;
-import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.moving;
-import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.movingUnique;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.blockingOrder;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.childRecordsCheck;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.newKeyNamed;
@@ -59,20 +51,19 @@ import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overridingThree;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.recordedChildBodyWithId;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.sourcing;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.streamMustInclude;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.withOpContext;
 import static com.hedera.services.bdd.suites.contract.hapi.ContractCallSuite.PAY_RECEIVABLE_CONTRACT;
 import static com.hedera.services.bdd.suites.contract.precompile.V1SecurityModelOverrides.CONTRACTS_MAX_NUM_WITH_HAPI_SIGS_ACCESS;
 import static com.hedera.services.bdd.suites.token.TokenAssociationSpecs.MULTI_KEY;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.CONTRACT_REVERT_EXECUTED;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_FULL_PREFIX_SIGNATURE_FOR_PRECOMPILE;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_SIGNATURE;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.NOT_SUPPORTED;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.SUCCESS;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.esaulpaugh.headlong.abi.Address;
 import com.google.protobuf.ByteString;
+import com.hedera.services.bdd.junit.HapiTest;
 import com.hedera.services.bdd.junit.HapiTestSuite;
 import com.hedera.services.bdd.spec.HapiSpec;
 import com.hedera.services.bdd.spec.HapiSpecOperation;
@@ -93,7 +84,7 @@ import org.junit.jupiter.api.Tag;
 public class ContractKeysStillWorkAsExpectedSuite extends HapiSuite {
 
     private static final Logger log = LogManager.getLogger(ContractKeysStillWorkAsExpectedSuite.class);
-    private static final String EVM_ALIAS_ENABLED_PROP = "cryptoCreateWithAliasAndEvmAddress.enabled";
+    public static final String EVM_ALIAS_ENABLED_PROP = "cryptoCreateWithAliasAndEvmAddress.enabled";
 
     public static void main(String... args) {
         new ContractKeysStillWorkAsExpectedSuite().runSuiteSync();
@@ -115,354 +106,10 @@ public class ContractKeysStillWorkAsExpectedSuite extends HapiSuite {
                 contractKeysStillHaveSpecificityNoMatterTopLevelSignatures(),
                 contractKeysWorkAsExpectedForFungibleTokenMgmt(),
                 canStillTransferByVirtueOfContractIdInEOAThreshold(),
-                approvalFallbacksRequiredWithoutTopLevelSigAccess(),
-                topLevelSigsStillWorkWithDefaultGrandfatherNum(),
-                contractCanStillTransferItsOwnAssets(),
-                fallbackFeeForHtsPayerMustSign(),
-                fallbackFeePayerMustSign(),
-                fixedFeeFailsWhenDisabledButWorksWhenEnabled());
+                approvalFallbacksRequiredWithoutTopLevelSigAccess());
     }
 
-    private HapiSpec fixedFeeFailsWhenDisabledButWorksWhenEnabled() {
-        final AtomicReference<Address> senderAddr = new AtomicReference<>();
-        final AtomicReference<Address> receiverAddr = new AtomicReference<>();
-        final AtomicReference<Address> nonFungibleTokenMirrorAddr = new AtomicReference<>();
-
-        final var treasury = "treasury";
-        final var sender = "sender";
-        final var receiver = "receiver";
-        final var nft = "nft";
-        final var failedCallTxn = "failedCallTxn";
-
-        return propertyPreservingHapiSpec("FixedFeeFailsWhenNotEnabled")
-                .preserving(EVM_ALIAS_ENABLED_PROP)
-                .given(
-                        overriding(EVM_ALIAS_ENABLED_PROP, "true"),
-                        uploadInitCode(WELL_KNOWN_TREASURY_CONTRACT),
-                        contractCreate(WELL_KNOWN_TREASURY_CONTRACT),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(treasury),
-                        cryptoCreate(sender).keyShape(SECP256K1_ON).exposingEvmAddressTo(senderAddr::set),
-                        getAccountInfo(sender).logged(),
-                        cryptoCreate(receiver)
-                                .keyShape(SECP256K1_ON)
-                                .balance(ONE_HBAR)
-                                .exposingEvmAddressTo(receiverAddr::set))
-                .when(
-                        tokenCreate(nft)
-                                .exposingAddressTo(nonFungibleTokenMirrorAddr::set)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(MULTI_KEY)
-                                .withCustom(fixedHbarFee(1, treasury, false))
-                                .treasury(treasury),
-                        getTokenInfo(nft).logged(),
-                        tokenAssociate(sender, nft),
-                        tokenAssociate(receiver, nft),
-                        mintToken(
-                                nft,
-                                List.of(ByteString.copyFromUtf8("serialNo1"), ByteString.copyFromUtf8("serialNo2"))),
-                        cryptoTransfer(movingUnique(nft, 1, 2).between(treasury, sender)))
-                .then(
-                        // override config to not allow fixed fees
-                        overriding("contracts.precompile.unsupportedCustomFeeReceiverDebits", "FIXED_FEE"),
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferSerialNo1FromToOthers",
-                                        nonFungibleTokenMirrorAddr.get(),
-                                        senderAddr.get(),
-                                        receiverAddr.get())
-                                .gas(2_000_000)
-                                .via(failedCallTxn)
-                                .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
-                                .alsoSigningWithFullPrefix(sender)),
-                        childRecordsCheck(
-                                failedCallTxn,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith().status(NOT_SUPPORTED)),
-                        // And no hbar were deducted from the receiver
-                        getAccountBalance(receiver).hasTinyBars(ONE_HBAR),
-                        // override config to  allow fixed fees
-                        overriding("contracts.precompile.unsupportedCustomFeeReceiverDebits", ""),
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferSerialNo1FromToOthers",
-                                        nonFungibleTokenMirrorAddr.get(),
-                                        senderAddr.get(),
-                                        receiverAddr.get())
-                                .gas(2_000_000)
-                                .via(failedCallTxn)
-                                .hasKnownStatus(SUCCESS)
-                                .alsoSigningWithFullPrefix(sender)));
-    }
-
-    private HapiSpec fallbackFeePayerMustSign() {
-        final AtomicReference<Address> senderAddr = new AtomicReference<>();
-        final AtomicReference<Address> receiverAddr = new AtomicReference<>();
-        final AtomicReference<Address> nonFungibleTokenMirrorAddr = new AtomicReference<>();
-
-        final var treasury = "treasury";
-        final var sender = "sender";
-        final var receiver = "receiver";
-        final var nft = "nft";
-        final var failedCallTxn = "failedCallTxn";
-
-        return propertyPreservingHapiSpec("FallbackFeePayerMustSign")
-                .preserving(EVM_ALIAS_ENABLED_PROP)
-                .given(
-                        overriding(EVM_ALIAS_ENABLED_PROP, "true"),
-                        uploadInitCode(WELL_KNOWN_TREASURY_CONTRACT),
-                        contractCreate(WELL_KNOWN_TREASURY_CONTRACT),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(treasury),
-                        cryptoCreate(sender).keyShape(SECP256K1_ON).exposingEvmAddressTo(senderAddr::set),
-                        getAccountInfo(sender).logged(),
-                        cryptoCreate(receiver)
-                                .keyShape(SECP256K1_ON)
-                                .balance(ONE_HBAR)
-                                .exposingEvmAddressTo(receiverAddr::set))
-                .when(
-                        tokenCreate(nft)
-                                .exposingAddressTo(nonFungibleTokenMirrorAddr::set)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(MULTI_KEY)
-                                .withCustom(royaltyFeeWithFallback(
-                                        1, 2, fixedHbarFeeInheritingRoyaltyCollector(ONE_HBAR), treasury))
-                                .treasury(treasury),
-                        getTokenInfo(nft).logged(),
-                        tokenAssociate(sender, nft),
-                        tokenAssociate(receiver, nft),
-                        mintToken(
-                                nft,
-                                List.of(ByteString.copyFromUtf8("serialNo1"), ByteString.copyFromUtf8("serialNo2"))),
-                        cryptoTransfer(movingUnique(nft, 1, 2).between(treasury, sender)))
-                .then(
-                        // Without the receiver signature, will fail with INVALID_SIGNATURE
-                        overriding("contracts.precompile.unsupportedCustomFeeReceiverDebits", ""),
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferSerialNo1FromToOthers",
-                                        nonFungibleTokenMirrorAddr.get(),
-                                        senderAddr.get(),
-                                        receiverAddr.get())
-                                .gas(2_000_000)
-                                .via(failedCallTxn)
-                                .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
-                                .alsoSigningWithFullPrefix(sender)),
-                        childRecordsCheck(
-                                failedCallTxn,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith().status(INVALID_FULL_PREFIX_SIGNATURE_FOR_PRECOMPILE)),
-                        // And no hbar were deducted from the receiver
-                        getAccountBalance(receiver).hasTinyBars(ONE_HBAR),
-                        // override config to not allow royalty fee
-                        overriding("contracts.precompile.unsupportedCustomFeeReceiverDebits", "ROYALTY_FALLBACK_FEE"),
-                        // But now sign with receiver as well
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferSerialNo1FromToOthers",
-                                        nonFungibleTokenMirrorAddr.get(),
-                                        senderAddr.get(),
-                                        receiverAddr.get())
-                                .gas(2_000_000)
-                                .via("removeRoyaltyFee")
-                                .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
-                                .alsoSigningWithFullPrefix(sender, receiver)),
-                        childRecordsCheck(
-                                "removeRoyaltyFee",
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith().status(NOT_SUPPORTED)),
-                        // And no hbar were deducted from the receiver
-                        getAccountBalance(receiver).hasTinyBars(ONE_HBAR),
-                        // But now sign with receiver as well and allow royalty fee
-                        overriding("contracts.precompile.unsupportedCustomFeeReceiverDebits", ""),
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferSerialNo1FromToOthers",
-                                        nonFungibleTokenMirrorAddr.get(),
-                                        senderAddr.get(),
-                                        receiverAddr.get())
-                                .gas(2_000_000)
-                                .alsoSigningWithFullPrefix(sender, receiver)),
-                        // Receiver balance will be debited ONE_HBAR
-                        getAccountBalance(receiver).hasTinyBars(0L));
-    }
-
-    private HapiSpec fallbackFeeForHtsPayerMustSign() {
-        final AtomicReference<Address> senderAddr = new AtomicReference<>();
-        final AtomicReference<Address> receiverAddr = new AtomicReference<>();
-        final AtomicReference<Address> nonFungibleTokenMirrorAddr = new AtomicReference<>();
-
-        final var treasury = "treasury";
-        final var sender = "sender";
-        final var receiver = "receiver";
-        final var nft = "nft";
-        final var fungible = "fungible";
-        final var failedCallTxn = "failedCallTxn";
-
-        return propertyPreservingHapiSpec("FallbackFeeHtsPayerMustSign")
-                .preserving(EVM_ALIAS_ENABLED_PROP)
-                .given(
-                        overriding(EVM_ALIAS_ENABLED_PROP, "true"),
-                        uploadInitCode(WELL_KNOWN_TREASURY_CONTRACT),
-                        contractCreate(WELL_KNOWN_TREASURY_CONTRACT),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(treasury),
-                        cryptoCreate(sender).keyShape(SECP256K1_ON).exposingEvmAddressTo(senderAddr::set),
-                        getAccountInfo(sender).logged(),
-                        cryptoCreate(receiver)
-                                .keyShape(SECP256K1_ON)
-                                .balance(ONE_HBAR)
-                                .exposingEvmAddressTo(receiverAddr::set))
-                .when(
-                        tokenCreate(fungible)
-                                .initialSupply(100)
-                                .tokenType(TokenType.FUNGIBLE_COMMON)
-                                .treasury(TOKEN_TREASURY),
-                        tokenAssociate(sender, fungible),
-                        tokenAssociate(receiver, fungible),
-                        cryptoTransfer(moving(10, fungible).between(TOKEN_TREASURY, receiver)),
-                        tokenCreate(nft)
-                                .exposingAddressTo(nonFungibleTokenMirrorAddr::set)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(MULTI_KEY)
-                                .withCustom(royaltyFeeWithFallback(
-                                        1, 2, fixedHtsFeeInheritingRoyaltyCollector(1, fungible), treasury))
-                                .treasury(treasury),
-                        getTokenInfo(nft).logged(),
-                        tokenAssociate(sender, nft),
-                        tokenAssociate(receiver, nft),
-                        mintToken(
-                                nft,
-                                List.of(ByteString.copyFromUtf8("serialNo1"), ByteString.copyFromUtf8("serialNo2"))),
-                        cryptoTransfer(movingUnique(nft, 1, 2).between(treasury, sender)))
-                .then(
-                        // Without the receiver signature, will fail with INVALID_SIGNATURE
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferSerialNo1FromToOthers",
-                                        nonFungibleTokenMirrorAddr.get(),
-                                        senderAddr.get(),
-                                        receiverAddr.get())
-                                .gas(2_000_000)
-                                .via(failedCallTxn)
-                                .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
-                                .alsoSigningWithFullPrefix(sender)),
-                        childRecordsCheck(
-                                failedCallTxn,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith().status(INVALID_FULL_PREFIX_SIGNATURE_FOR_PRECOMPILE)),
-                        // And no tokens were deducted from the receiver
-                        getAccountBalance(receiver).hasTokenBalance(fungible, 10),
-                        // But now sign with receiver as well
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferSerialNo1FromToOthers",
-                                        nonFungibleTokenMirrorAddr.get(),
-                                        senderAddr.get(),
-                                        receiverAddr.get())
-                                .gas(2_000_000)
-                                .alsoSigningWithFullPrefix(sender, receiver)),
-                        // Receiver token balance of custom fee denomination should debit by 1
-                        getAccountBalance(receiver).hasTokenBalance(fungible, 9));
-    }
-
-    private HapiSpec contractCanStillTransferItsOwnAssets() {
-        final AtomicReference<Address> fungibleTokenMirrorAddr = new AtomicReference<>();
-        final AtomicReference<Address> nonFungibleTokenMirrorAddr = new AtomicReference<>();
-        final AtomicReference<Address> aSenderAddr = new AtomicReference<>();
-        final AtomicReference<Address> aReceiverAddr = new AtomicReference<>();
-        final AtomicReference<Address> bSenderAddr = new AtomicReference<>();
-        final AtomicReference<Address> bReceiverAddr = new AtomicReference<>();
-        final AtomicReference<Address> treasuryContractAddr = new AtomicReference<>();
-
-        return defaultHapiSpec("ContractCanStillTransferItsOwnAssets")
-                .given(someWellKnownTokensAndAccounts(
-                        fungibleTokenMirrorAddr,
-                        nonFungibleTokenMirrorAddr,
-                        aSenderAddr,
-                        aReceiverAddr,
-                        bSenderAddr,
-                        bReceiverAddr,
-                        false))
-                .when(withOpContext((spec, opLog) -> {
-                    final var treasuryId = spec.registry().getAccountID(WELL_KNOWN_TREASURY_CONTRACT);
-                    treasuryContractAddr.set(idAsHeadlongAddress(treasuryId));
-                }))
-                .then(
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        TOKEN_UNIT_FROM_TO_OTHERS_TXN,
-                                        fungibleTokenMirrorAddr.get(),
-                                        treasuryContractAddr.get(),
-                                        aReceiverAddr.get())
-                                .gas(2_000_000)),
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferSerialNo1FromToOthers",
-                                        nonFungibleTokenMirrorAddr.get(),
-                                        treasuryContractAddr.get(),
-                                        aReceiverAddr.get())
-                                .gas(2_000_000)),
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferNFTSerialNos2And3ToFromToOthers",
-                                        nonFungibleTokenMirrorAddr.get(),
-                                        treasuryContractAddr.get(),
-                                        aReceiverAddr.get(),
-                                        bSenderAddr.get(),
-                                        bReceiverAddr.get())
-                                .gas(2_000_000)
-                                .alsoSigningWithFullPrefix(B_WELL_KNOWN_SENDER)),
-                        sourcing(() -> contractCall(
-                                        WELL_KNOWN_TREASURY_CONTRACT,
-                                        "transferTokenUnitsFromToOthers",
-                                        fungibleTokenMirrorAddr.get(),
-                                        treasuryContractAddr.get(),
-                                        aReceiverAddr.get(),
-                                        bSenderAddr.get(),
-                                        bReceiverAddr.get())
-                                .gas(2_000_000)
-                                .alsoSigningWithFullPrefix(B_WELL_KNOWN_SENDER)),
-                        someWellKnownAssertions());
-    }
-
-    private HapiSpec topLevelSigsStillWorkWithDefaultGrandfatherNum() {
-        final AtomicReference<Address> fungibleTokenMirrorAddr = new AtomicReference<>();
-        final AtomicReference<Address> nonFungibleTokenMirrorAddr = new AtomicReference<>();
-        final AtomicReference<Address> aSenderAddr = new AtomicReference<>();
-        final AtomicReference<Address> aReceiverAddr = new AtomicReference<>();
-        final AtomicReference<Address> bSenderAddr = new AtomicReference<>();
-        final AtomicReference<Address> bReceiverAddr = new AtomicReference<>();
-
-        return defaultHapiSpec("TopLevelSigsStillWorkWithDefaultGrandfatherNum")
-                .given(
-                        // Top-level signatures are available to any contract with 10M grandfather
-                        // num
-                        someWellKnownTokensAndAccounts(
-                                fungibleTokenMirrorAddr,
-                                nonFungibleTokenMirrorAddr,
-                                aSenderAddr,
-                                aReceiverAddr,
-                                bSenderAddr,
-                                bReceiverAddr,
-                                true))
-                .when()
-                .then(
-                        // Everything should have the needed approvals
-                        someWellKnownOperationsWithAllNeededSigsInSigMap(
-                                fungibleTokenMirrorAddr,
-                                nonFungibleTokenMirrorAddr,
-                                aSenderAddr,
-                                aReceiverAddr,
-                                bSenderAddr,
-                                bReceiverAddr,
-                                SUCCESS),
-                        someWellKnownAssertions());
-    }
-
+    // No such record stream file location: <mono-service>
     private HapiSpec approvalFallbacksRequiredWithoutTopLevelSigAccess() {
         final AtomicReference<Address> fungibleTokenMirrorAddr = new AtomicReference<>();
         final AtomicReference<Address> nonFungibleTokenMirrorAddr = new AtomicReference<>();
@@ -548,14 +195,14 @@ public class ContractKeysStillWorkAsExpectedSuite extends HapiSuite {
     }
 
     private static final String A_WELL_KNOWN_SENDER = "A_SENDER";
-    private static final String B_WELL_KNOWN_SENDER = "B_SENDER";
+    public static final String B_WELL_KNOWN_SENDER = "B_SENDER";
     private static final String A_WELL_KNOWN_RECEIVER = "A_RECEIVER";
     private static final String B_WELL_KNOWN_RECEIVER = "B_RECEIVER";
     private static final String WELL_KNOWN_FUNGIBLE_TOKEN = "FT";
     private static final String WELL_KNOWN_NON_FUNGIBLE_TOKEN = "NFT";
     private static final String WELL_KNOWN_TREASURY_CONTRACT = "DoTokenManagement";
 
-    private HapiSpecOperation someWellKnownAssertions() {
+    public static HapiSpecOperation someWellKnownAssertions() {
         return blockingOrder(
                 getAccountBalance(A_WELL_KNOWN_RECEIVER)
                         .hasTokenBalance(WELL_KNOWN_FUNGIBLE_TOKEN, 2L)
@@ -573,7 +220,7 @@ public class ContractKeysStillWorkAsExpectedSuite extends HapiSuite {
      * that serves as both their treasuries; and some accounts associated to both tokens, where the
      * sender accounts have fungible units and serial numbers 1 and 2, respectively.
      */
-    private HapiSpecOperation someWellKnownTokensAndAccounts(
+    public static HapiSpecOperation someWellKnownTokensAndAccounts(
             final AtomicReference<Address> fungibleTokenMirrorAddr,
             final AtomicReference<Address> nonFungibleTokenMirrorAddr,
             final AtomicReference<Address> aSenderAddr,
@@ -629,14 +276,14 @@ public class ContractKeysStillWorkAsExpectedSuite extends HapiSuite {
                 contractUpdate(WELL_KNOWN_TREASURY_CONTRACT).properlyEmptyingAdminKey());
     }
 
-    private static final String TOKEN_UNIT_FROM_TO_OTHERS_TXN = "transferTokenUnitFromToOthers";
+    public static final String TOKEN_UNIT_FROM_TO_OTHERS_TXN = "transferTokenUnitFromToOthers";
 
     /**
      * Returns a multi-step operation that does one of each of the {@code transferToken}, {@code
      * transferTokens}, {@code transferNFT}, {@code transferNFTs} with the given expected status.
      * Every operation adds all the needed signatures to the sigMap using full-prefix signatures.
      */
-    private HapiSpecOperation someWellKnownOperationsWithAllNeededSigsInSigMap(
+    public static HapiSpecOperation someWellKnownOperationsWithAllNeededSigsInSigMap(
             final AtomicReference<Address> fungibleTokenMirrorAddr,
             final AtomicReference<Address> nonFungibleTokenMirrorAddr,
             final AtomicReference<Address> aSenderAddr,
@@ -688,6 +335,7 @@ public class ContractKeysStillWorkAsExpectedSuite extends HapiSuite {
                         .hasKnownStatus(expectedStatus)));
     }
 
+    @HapiTest
     private HapiSpec canStillTransferByVirtueOfContractIdInEOAThreshold() {
         final var fungibleToken = "token";
         final var managementContract = "DoTokenManagement";
@@ -750,6 +398,7 @@ public class ContractKeysStillWorkAsExpectedSuite extends HapiSuite {
                         getAccountBalance(receiver).logged().hasTokenBalance(fungibleToken, 1));
     }
 
+    // Expected a contract admin key, got ed25519: "..."
     private HapiSpec contractKeysWorkAsExpectedForFungibleTokenMgmt() {
         final var fungibleToken = "token";
         final var managementContract = "DoTokenManagement";
@@ -812,6 +461,7 @@ public class ContractKeysStillWorkAsExpectedSuite extends HapiSuite {
                                 .logged());
     }
 
+    // Expected a contract admin key, got ed25519: "..."
     private HapiSpec contractKeysStillHaveSpecificityNoMatterTopLevelSignatures() {
         final var fungibleToken = "token";
         final var managementContract = "DoTokenManagement";
