+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/precompile/ERCPrecompileSuite.java
@@ -3065,6 +3065,7 @@ public class ERCPrecompileSuite extends HapiSuite {
                                                         .withAllowance(0)))));
     }
 
+    @HapiTest
     private HapiSpec erc20TransferFromSelf() {
         return defaultHapiSpec("erc20TransferFromSelf")
                 .given(
